Clean up todos methods
